elf-loader: rename app and loader
