+++ applications/elf-loader/elf-lib/elf-loader.h
